+++ liferay-plugins-sdk-6.2/portlets/flask-user-tailgate-portlet/docroot/WEB-INF/service/com/rumbasolutions/flask/service/TailgateInfoServiceWrapper.java
@@ -60,7 +60,7 @@ public class TailgateInfoServiceWrapper implements TailgateInfoService,
 	public com.rumbasolutions.flask.model.TailgateInfo addTailgateInfo(
 		java.lang.String tailgateName, java.lang.String tailgateDescription,
 		long eventId, java.lang.String eventName, java.util.Date tailgateDate,
-		java.util.Date startTime, java.util.Date endTime, double amountToPay,
+		java.util.Date startTime, java.util.Date endTime, long amountToPay,
 		com.liferay.portal.service.ServiceContext serviceContext) {
 		return _tailgateInfoService.addTailgateInfo(tailgateName,
 			tailgateDescription, eventId, eventName, tailgateDate, startTime,
@@ -89,7 +89,7 @@ public class TailgateInfoServiceWrapper implements TailgateInfoService,
 		long tailgateId, java.lang.String tailgateName,
 		java.lang.String tailgateDescription, long eventId,
 		java.lang.String eventName, java.util.Date tailgateDate,
-		java.util.Date startTime, java.util.Date endTime, double amountToPay,
+		java.util.Date startTime, java.util.Date endTime, long amountToPay,
 		com.liferay.portal.service.ServiceContext serviceContext) {
 		return _tailgateInfoService.updateTailgateInfo(tailgateId,
 			tailgateName, tailgateDescription, eventId, eventName,
