+++ liferay-plugins-sdk-6.2/portlets/flask-user-tailgate-portlet/docroot/WEB-INF/sql/tables.sql
@@ -25,7 +25,7 @@ create table flasktailgate_TailgateInfo (
 	endTime DATE null,
 	isActive INTEGER,
 	isDelete INTEGER,
-	amountToPay DOUBLE
+	amountToPay LONG
 );
 
 create table flasktailgate_TailgateMarker (
